package/linux-firmware: exclude some files from compression
Turns out some files in the firmware folder are not loaded by the firmware
loader mechanism, which results in failure to find those files when a module
wants them. So far only iwlwifi's .pnvm files seem to be this case, but added a
universal logic that can be used for excluding individual files or glob
patterns.