package/linux-firmware: fix AX210 support, stick to latest supported ucode API, reorganize the split
Adjust the hardcoded ucode API versions for Intel cards to use constants
defined in kernel driver definitions. This will make maintaining this
downstream quirk a little more bearable.

When doing this, I noticed the IWLWIFI_6E option is actually missing firmware
for AX210 (ty-a0-gf-a0) which was provided through IWLWIFI_22000, but there it
was missing the pnvm file (so-a0-gf-a0 was for AX211). Also the IWLWIFI_22260
option was doing nothing, so I moved the correct firmwares from 22000 there.

Signed-off-by: Jan Čermák <sairon@sairon.cz>